@@ -44,31 +44,31 @@
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
-    "@types/benchmark": "^1.0.31",
-    "@types/fs-extra": "^8.0.1",
-    "@types/jest": "^24.0.20",
-    "@types/json2csv": "^4.5.0",
-    "@types/node": "^12.11.7",
-    "@types/prettier": "^1.18.3",
+    "@types/benchmark": "^2.1.0",
+    "@types/fs-extra": "^9.0.4",
+    "@types/jest": "^26.0.15",
+    "@types/json2csv": "^5.0.1",
+    "@types/node": "^14.14.7",
+    "@types/prettier": "^2.1.5",
     "codecov": "^3.6.1",
     "csvtojson": "^2.0.10",
-    "jest": "^24.9.0",
-    "jsdom": "^15.2.0",
-    "ts-jest": "^24.1.0",
-    "tslint": "^5.20.0",
+    "jest": "^26.6.3",
+    "jsdom": "^16.4.0",
+    "ts-jest": "^26.4.4",
+    "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "^3.6.3"
+    "typescript": "^4.0.5"
   },
   "dependencies": {
     "@arrows/composition": "^1.0.0",
     "@arrows/dispatch": "^1.0.2",
     "@arrows/multimethod": "^1.1.6",
     "benchmark": "^2.1.4",
-    "fs-extra": "^8.1.0",
-    "json2csv": "^4.5.3",
-    "kleur": "^3.0.3",
-    "log-update": "^3.3.0",
-    "prettier": "^1.18.2",
+    "fs-extra": "^9.0.1",
+    "json2csv": "^5.0.4",
+    "kleur": "^4.1.3",
+    "log-update": "^4.0.0",
+    "prettier": "^2.1.2",
     "stats-median": "^1.0.1"
   }
 }
